Initial commit, repo scaffold
I'm going to try Advent of Code in Clojure this year. I've never really
learned Clojure, so this should be fun!

I'm using Leinengen as that seems to be the equivalent of Maven in the
Java world.

Since I know AoC usually takes the form of reading an input file, I've
implemented a basic "cat" clone that reads from stdin.